refactor(package.js): separate functions into different file
